refactor(database): decompose the database core god-file by role (WS-D)
Split internal/database/database.go (684 LOC) into role-based files in the
same package — pure move: DB type, config, open/corruption-recovery, and
connection lifecycle stay in database.go; the per-repository accessor
methods move to repositories.go.

All files under the 600-line size-gate threshold. Declaration sets and
code-line multisets verified identical to the original.

Related to #187
